Many updates made
Project is completed! (Kind of)
Functions off of UDP connections only.
Server messages are displayed in a JFrame for the client so that system messages do not interrupt user input.

Next step is to make bid work over TCP.
Also need to finish commenting on code.
Some slight optimizations can be made (consistency).

There is also a bug when restoring from a backup. Clients can still send messages however the server fails to return the messages to the client... Will work on this later.